Instrumentation DynamoDB AWS SDK v3 [skip ci]
