+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- AWS SDK v3: Instrumentation of DynamoDB
+
 ## 1.126.2
 - Fix: AWS SQS headers reading properly checks for strings, numbers and arrays of strings
