+++ package.json
@@ -34,6 +34,7 @@
   "devDependencies": {
     "@apollo/federation": "^0.25.1",
     "@apollo/gateway": "^0.28.2",
+    "@aws-sdk/client-dynamodb": "^3.20.0",
     "@elastic/elasticsearch": "^7.12.0",
     "@google-cloud/pubsub": "^2.12.0",
     "@google-cloud/storage": "^5.8.5",
