+++ packages/collector/test/tracing/cloud/aws-sdk/v2/dynamodb/util.js
