+++ packages/collector/test/tracing/cloud/aws-sdk/v2/kinesis/util.js
