+++ packages/collector/test/tracing/cloud/aws-sdk/v2/sqs/sendNonInstrumented.js
