+++ packages/collector/test/tracing/cloud/aws-sdk/v2/sqs/sqsUtil.js
