+++ packages/core/src/tracing/index.js
@@ -30,9 +30,10 @@ let processIdentityProvider = null;
 
 // Note: Also update initializedTooLateHeuristic.js and the accompanying test when adding instrumentations.
 const instrumentations = [
-  './instrumentation/cloud/aws/index',
-  './instrumentation/cloud/aws/sdk',
-  './instrumentation/cloud/aws/sqs',
+  './instrumentation/cloud/aws-sdk/v2/index',
+  './instrumentation/cloud/aws-sdk/v3/index',
+  './instrumentation/cloud/aws-sdk/v2/sdk',
+  './instrumentation/cloud/aws-sdk/v2/sqs',
   './instrumentation/cloud/gcp/pubsub',
   './instrumentation/cloud/gcp/storage',
   './instrumentation/control_flow/bluebird',
