+++ packages/core/src/tracing/instrumentation/cloud/aws-sdk/v2/instana_aws_product.js
